stable version 0.5

+++ include/trival_helper.h
@@ -16,6 +16,6 @@ void print_time(struct tm* lt);
 bool test_bit_char(char byte, int num);
 char fill_bit_char(char byte, int num);
 char release_bit_char(char byte, int num);
-unsigned int priority_get_by_number(int num);
+unsigned long priority_get_by_number(int num);
 
 #endif //VFS_TRIVAL_HELPER_H
